Introduces CMake, improves the README, and restructures project
The old set up got a little stale. Due to not having a define build tool, it became difficult to set up the project after switching to a new computer. In an effort to make this more easily runnable by anyone, I introduced CMake to improve the set up process.

A description of the project is now in the README.

While adding CMake, I separated the headers from the source code and restructured the project. Minimal code changes were made. The primary changes were with the `#include` sections.